refactor: separate memo generation for pdt (v1 & v2) and vaccination
